Fixing SQLRequest
@victorsmits

+++ CodeSource/KitBoxSourceCode/KitBoxSourceCode/Cabinet.cs
@@ -9,9 +9,9 @@ namespace KitBoxSourceCode
         private List<CabinetFloor> cabinetFloors;
         private int cabinetHeight;
 
-        private int cabinetPrice = 0;
+        private double cabinetPrice = 0;
 
-        public int GetCabinetPrice => cabinetPrice;
+        public double GetCabinetPrice => cabinetPrice;
         public int GetCabinetHeight => cabinetHeight;
 
         private Angle angles;
